update relations exception handle - center in InjectionManager
