+++ database/inject/Injectable.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+from collections import defaultdict
 
 class Injectable(ABC):
     @classmethod
@@ -13,3 +14,12 @@ class Injectable(ABC):
     async def process(cls, replay, session):
         """Must be implemented by subclasses."""
         pass
+
+    @classmethod
+    def get_data(cls, obj):
+        parameters = defaultdict(lambda:None)
+        for variable, value in vars(obj).items():
+            if variable in cls.columns:
+                parameters[variable] = value
+
+        return parameters
