+++ database/warehouse/replay/map.py
@@ -3,14 +3,11 @@ from sqlalchemy.exc import IntegrityError, OperationalError
 from sqlalchemy.future import select
 from sqlalchemy.orm import relationship
 
-from asyncio import Lock
-
 from database.inject import Injectable
 from database.base import Base
 
 
 class map(Injectable, Base):
-    _lock = Lock()
     __tablename__ = "map"
     __table_args__ = {"schema": "replay"}
 
@@ -22,7 +19,6 @@ class map(Injectable, Base):
     author = Column(Text)
     description = Column(Text)
     website = Column(Text)
-    minimap = Column(LargeBinary) ## Store elsewhere?
 
     replays = relationship("info", back_populates="map")
 
@@ -33,43 +29,20 @@ class map(Injectable, Base):
 
     @classmethod
     async def process(cls, replay, session):
-        async with cls._lock:
-            try:
-                if await cls.process_existence(replay.map_hash, session):
-                    return
+        if await cls.process_existence(replay.map_hash, session):
+            return
 
-                # Load map if not exists
-                replay.load_map()
+        # Load map if not exists
+        replay.load_map()
 
-                data = cls.get_data(replay.map)
-                session.add(cls(**data))
-
-            except IntegrityError as e:
-                await session.rollback()
-                print(f"IntegrityError: {e.orig}")
-                # Handle specific cases like unique constraint violations
-            except OperationalError as e:
-                await session.rollback()
-                print(f"OperationalError: {e.orig}")
-                # Handle deadlocks or connection issues
-            except Exception as e:
-                await session.rollback()
-                print(f"Unexpected error: {e}")
-                # Gracefully handle all other exceptions
+        data = cls.get_data(replay.map)
+        session.add(cls(**data))
 
     @classmethod
     async def process_existence(cls, filehash, session):
         statement = select(cls).where(cls.filehash == filehash)
         result = await session.execute(statement)
         return result.scalar()
-
-    @classmethod
-    def get_data(cls, obj):
-        parameters = {}
-        for variable, value in vars(obj).items():
-            if variable in cls.columns:
-                parameters[variable] = value
-        return parameters
 
     columns = \
         { "filename"
@@ -78,5 +51,4 @@ class map(Injectable, Base):
         , "author"
         , "description"
         , "website"
-        , "minimap"
         }
